Remove Get a Job section
Including 2.5 pages of content on finding a job elsewhere does not feel
like the right message to send to new apprentices joining the thoughtbot
team. It is all good content, but should be moved elsewhere.

Also, updating section about the "end of the second month" conversation
to reference myself for San Francisco apprentices and include more
details about how that conversation goes down.