Update onboarding
This was out of date or incorrect in several areas. I updated it some,
and it could likely use more attention and verification that the
suggested activities are still valid.

* Used `apprenticeship` rather than `apprentice.io`. We haven't been
using the `apprentice.io` branding of late, so introducing apprentices
to that seems more confusing than helpful.
* Switched to markdown style links. It seems this document was
originally intended to be printed. I did that for someone once. It seems
like it's much more valuable as a resource we'd give someone
electronically (with clickable links!)
* Removed some unnecessarily strong or assumptive language (e.g.
"Weapons", "Meeting people is easy!").
* Added section on Constable
* Removed section on Basecamp.
* Fixed various other errors/omissions.