Apprentices are already hired
Since this was originally written we have significantly refined the apprenticeship at thoughtbot.

When you are hired as an Apprentice at thoughtbot it is a full-time position with full benefits. There is no question whether you will be "hired" by thoughtbot at the end of the apprenticeship, because you have already been hired.
Rather you are working to be promoted to a full designer or developer.

Like with any job, there are rare circumstances where a position doesn't work out and we need to part ways, but that is the exception, and everything is oriented toward working on promotion from your current role.